+++ pkg/config/config_test.go
@@ -1,12 +1,12 @@
-package cmd
+package config_test
 
 import (
 	"fmt"
-	"os"
-
+	"github.com/bakito/adguardhome-sync/pkg/config"
 	"github.com/bakito/adguardhome-sync/pkg/types"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
+	"os"
 )
 
 var envVars = []string{
@@ -55,7 +55,7 @@ var _ = Describe("Run", func() {
 		})
 		Context("getConfig", func() {
 			It("features should be false", func() {
-				cfg, err := getConfig()
+				cfg, err := config.Get("")
 				Ω(err).ShouldNot(HaveOccurred())
 				verifyFeatures(cfg, false)
 			})
@@ -74,12 +74,12 @@ var _ = Describe("Run", func() {
 		})
 		Context("getConfig", func() {
 			It("features should be true by default", func() {
-				cfg, err := getConfig()
+				cfg, err := config.Get("")
 				Ω(err).ShouldNot(HaveOccurred())
 				verifyFeatures(cfg, true)
 			})
 			It("features should be true by default", func() {
-				cfg, err := getConfig()
+				cfg, err := config.Get("")
 				Ω(err).ShouldNot(HaveOccurred())
 				verifyFeatures(cfg, true)
 			})
@@ -87,7 +87,7 @@ var _ = Describe("Run", func() {
 				for _, envVar := range envVars {
 					Ω(os.Setenv(envVar, "false")).ShouldNot(HaveOccurred())
 				}
-				cfg, err := getConfig()
+				cfg, err := config.Get("")
 				Ω(err).ShouldNot(HaveOccurred())
 				verifyFeatures(cfg, false)
 			})
@@ -95,7 +95,7 @@ var _ = Describe("Run", func() {
 				It("should set interface name of replica 1", func() {
 					Ω(os.Setenv("REPLICA1_URL", "https://foo.bar")).ShouldNot(HaveOccurred())
 					Ω(os.Setenv(fmt.Sprintf("REPLICA%s_INTERFACE_NAME", "1"), "eth0")).ShouldNot(HaveOccurred())
-					cfg, err := getConfig()
+					cfg, err := config.Get("")
 					Ω(err).ShouldNot(HaveOccurred())
 					Ω(cfg.Replicas[0].InterfaceName).Should(Equal("eth0"))
 				})
@@ -104,7 +104,7 @@ var _ = Describe("Run", func() {
 				It("should enable the dhcp server of replica 1", func() {
 					Ω(os.Setenv("REPLICA1_URL", "https://foo.bar")).ShouldNot(HaveOccurred())
 					Ω(os.Setenv(fmt.Sprintf("REPLICA%s_DHCPSERVERENABLED", "1"), "true")).ShouldNot(HaveOccurred())
-					cfg, err := getConfig()
+					cfg, err := config.Get("")
 					Ω(err).ShouldNot(HaveOccurred())
 					Ω(cfg.Replicas[0].DHCPServerEnabled).ShouldNot(BeNil())
 					Ω(*cfg.Replicas[0].DHCPServerEnabled).Should(BeTrue())
@@ -112,7 +112,7 @@ var _ = Describe("Run", func() {
 				It("should disable the dhcp server of replica 1", func() {
 					Ω(os.Setenv("REPLICA1_URL", "https://foo.bar")).ShouldNot(HaveOccurred())
 					Ω(os.Setenv(fmt.Sprintf("REPLICA%s_DHCPSERVERENABLED", "1"), "false")).ShouldNot(HaveOccurred())
-					cfg, err := getConfig()
+					cfg, err := config.Get("")
 					Ω(err).ShouldNot(HaveOccurred())
 					Ω(cfg.Replicas[0].DHCPServerEnabled).ShouldNot(BeNil())
 					Ω(*cfg.Replicas[0].DHCPServerEnabled).Should(BeFalse())
